cloud/requirements.txt
paho-mqtt==1.6.1
kafka-python==2.0.1
pandas==1.5.1
influxdb-client==1.35.0